Move predefined task model
