+++ src/azure/devops/workflow/pre-defined-tasks/development/setup-environment.ts
@@ -1,4 +1,4 @@
-import { PreDefinedTask } from "../../pre-defined-task";
+import { PreDefinedTask } from "../pre-defined-task";
 
 /**
  * Represents a pre-defined task for setting up the development environment.
