+++ src/azure/devops/workflow/pre-defined-tasks/planning/create-implementation-tasks.ts
@@ -1,4 +1,4 @@
-import { PreDefinedTask } from "../../pre-defined-task";
+import { PreDefinedTask } from "../pre-defined-task";
 
 /**
  * Represents a pre-defined task to create tasks for implementing the changes needed to meet a work item's acceptance criteria.
