+++ src/azure/devops/workflow/pre-defined-tasks/planning/meet-with-product-owners.ts
@@ -1,4 +1,4 @@
-import { PreDefinedTask } from "../../pre-defined-task";
+import { PreDefinedTask } from "../pre-defined-task";
 
 /**
  * Represents a pre-defined task for meeting with product owners to ensure alignment with business objectives.
