+++ src/azure/devops/workflow/pre-defined-tasks/planning/meet-with-stakeholders.ts
@@ -1,4 +1,4 @@
-import { PreDefinedTask } from "../../pre-defined-task";
+import { PreDefinedTask } from "../pre-defined-task";
 
 /**
  * Represents a pre-defined task for meeting with stakeholders to gather feedback and ensure alignment with expectations.
