+++ src/azure/devops/workflow/pre-defined-tasks/planning/reproduce-locally.ts
@@ -1,4 +1,4 @@
-import { PreDefinedTask } from "../../pre-defined-task";
+import { PreDefinedTask } from "../pre-defined-task";
 import { WorkItemType } from "../../../work-item-type";
 
 /**
