+++ src/azure/devops/workflow/pre-defined-tasks/planning/review-existing-implementations.ts
@@ -1,4 +1,4 @@
-import { PreDefinedTask } from "../../pre-defined-task";
+import { PreDefinedTask } from "../pre-defined-task";
 
 /**
  * Represents a pre-defined task for reviewing existing implementations to gather insights and derive potential solutions.
