+++ src/azure/devops/workflow/pre-defined-tasks/planning/review-pre-defined-tasks.ts
@@ -1,4 +1,4 @@
-import { PreDefinedTask } from "../../pre-defined-task";
+import { PreDefinedTask } from "../pre-defined-task";
 
 /**
  * Represents a pre-defined task for ensuring that pre-defined tasks are aligned with the objectives of the work item and that the remaining work for each task is accurate.
