+++ src/azure/devops/workflow/pre-defined-tasks/quality-assurance/create-qa-build.ts
@@ -1,4 +1,4 @@
-import { PreDefinedTask } from "../../pre-defined-task";
+import { PreDefinedTask } from "../pre-defined-task";
 
 /**
  * Represents a pre-defined task for creating a quality assurance build in Azure DevOps.
