+++ src/azure/devops/workflow/pre-defined-tasks/quality-assurance/dev-review-test-cases.ts
@@ -1,4 +1,4 @@
-import { PreDefinedTask } from "../../pre-defined-task";
+import { PreDefinedTask } from "../pre-defined-task";
 
 /**
  * Represents a pre-defined task for the developer review of test cases in Azure DevOps.
