+++ src/azure/devops/workflow/pre-defined-tasks/quality-assurance/qa-smoke-testing.ts
@@ -1,4 +1,4 @@
-import { PreDefinedTask } from "../../pre-defined-task";
+import { PreDefinedTask } from "../pre-defined-task";
 
 /**
  * Represents a pre-defined task for smoke testing in Azure DevOps.
