+++ src/azure/devops/workflow/pre-defined-tasks/quality-assurance/qa-test-case-execution.ts
@@ -1,4 +1,4 @@
-import { PreDefinedTask } from "../../pre-defined-task";
+import { PreDefinedTask } from "../pre-defined-task";
 
 /**
  * Represents a pre-defined task for executing test cases in Azure DevOps.
