+++ src/azure/devops/workflow/pre-defined-tasks/quality-assurance/qa-write-test-cases.ts
@@ -1,4 +1,4 @@
-import { PreDefinedTask } from "../../pre-defined-task";
+import { PreDefinedTask } from "../pre-defined-task";
 
 /**
  * Represents a pre-defined task for writing test cases in Azure DevOps.
