+++ src/azure/devops/workflow/pre-defined-tasks/quality-assurance/support-qa-testing.ts
@@ -1,4 +1,4 @@
-import { PreDefinedTask } from "../../pre-defined-task";
+import { PreDefinedTask } from "../pre-defined-task";
 
 /**
  * Represents a pre-defined task to support the QA team during the testing process in Azure DevOps.
