+++ src/azure/devops/workflow/pre-defined-tasks/release/create-release-build.ts
@@ -1,4 +1,4 @@
-import { PreDefinedTask } from "../../pre-defined-task";
+import { PreDefinedTask } from "../pre-defined-task";
 
 /**
  * Represents a pre-defined task for creating a release build in Azure DevOps.
