+++ src/azure/devops/workflow/pre-defined-tasks/release/create-release-notes.ts
@@ -1,4 +1,4 @@
-import { PreDefinedTask } from "../../pre-defined-task";
+import { PreDefinedTask } from "../pre-defined-task";
 
 /**
  * Represents a pre-defined task to create release notes in Azure DevOps.
