+++ src/azure/devops/workflow/pre-defined-tasks/release/deploy-to-production.ts
@@ -1,4 +1,4 @@
-import { PreDefinedTask } from "../../pre-defined-task";
+import { PreDefinedTask } from "../pre-defined-task";
 
 /**
  * Represents a pre-defined task for deploying the release build to the production environment in Azure DevOps.
