+++ src/azure/devops/workflow/pre-defined-tasks/release/notify-stakeholders-of-deployment.ts
@@ -1,4 +1,4 @@
-import { PreDefinedTask } from "../../pre-defined-task";
+import { PreDefinedTask } from "../pre-defined-task";
 
 /**
  * Represents a pre-defined task for notifying stakeholders about the deployment to production in Azure DevOps.
