+++ src/azure/devops/workflow/pre-defined-tasks/release/validate-in-production.ts
@@ -1,4 +1,4 @@
-import { PreDefinedTask } from "../../pre-defined-task";
+import { PreDefinedTask } from "../pre-defined-task";
 
 /**
  * Represents a pre-defined task for validating the functionality of the deployed build in the production environment in Azure DevOps.
