+++ src/azure/devops/workflow/pre-defined-tasks/review/complete-pr.ts
@@ -1,4 +1,4 @@
-import { PreDefinedTask } from "../../pre-defined-task";
+import { PreDefinedTask } from "../pre-defined-task";
 
 /**
  * Represents a pre-defined task to complete the pull request after all validations and reviews are successful in Azure DevOps.
