+++ src/azure/devops/workflow/pre-defined-tasks/review/create-draft-pr.ts
@@ -1,4 +1,4 @@
-import { PreDefinedTask } from "../../pre-defined-task";
+import { PreDefinedTask } from "../pre-defined-task";
 
 /**
  * Represents a pre-defined task for creating draft pull request in Azure DevOps.
