+++ src/azure/devops/workflow/pre-defined-tasks/review/resolve-pr-feedback.ts
@@ -1,4 +1,4 @@
-import { PreDefinedTask } from "../../pre-defined-task";
+import { PreDefinedTask } from "../pre-defined-task";
 
 /**
  * Represents a pre-defined task for resolving feedback provided during the pull request review process in Azure DevOps.
