+++ src/azure/devops/workflow/pre-defined-tasks/review/run-pr-validations.ts
@@ -1,4 +1,4 @@
-import { PreDefinedTask } from "../../pre-defined-task";
+import { PreDefinedTask } from "../pre-defined-task";
 
 /**
  * Represents a pre-defined task for running validations for the pull request in Azure DevOps.
