+++ src/azure/devops/workflow/pre-defined-tasks/review/self-review-pr.ts
@@ -1,4 +1,4 @@
-import { PreDefinedTask } from "../../pre-defined-task";
+import { PreDefinedTask } from "../pre-defined-task";
 
 /**
  * Represents a pre-defined task for self-reviewing a pull request in Azure DevOps.
